voluncheer/requirements.txt
#Put requirements here to have a consistent environment for everyone if necessary.
django-debug-toolbar